requirements.txt
python-dotenv>=1.0.0
gradio>=4.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-chroma>=0.1.0
chromadb>=0.4.0
openai>=1.0.0
